Refactore minimizer files
